Strategy implemented, working first pass of basehandlerid
