+++ src/daqpytools/logging/handlers.py
@@ -3,323 +3,38 @@ from __future__ import annotations
 import copy
 import io
 import logging
-import os
-import re
 import sys
 import time
 from collections import defaultdict
 from collections.abc import Callable
-from dataclasses import dataclass, field
 from datetime import datetime
-from enum import Enum
 from threading import Lock
-from typing import ClassVar, cast
+from typing import cast
 
 from erskafka.ERSKafkaLogHandler import ERSKafkaLogHandler
-from rich.console import Console, ConsoleRenderable
 from rich.logging import RichHandler
 from rich.text import Text
 
 from daqpytools.logging.exceptions import (
-    ERSEnvError,
     LoggerHandlerError,
-    ProtobufFormatError,
 )
 from daqpytools.logging.formatter import (
-    CONSOLE_THEME,
     DATE_TIME_BASE_FORMAT,
-    DATE_TIME_FORMAT,
     LOG_RECORD_PADDING,
     TIME_ZONE,
     LoggingFormatter,
 )
-from daqpytools.logging.levels import level_to_ers_var, logging_log_level_to_str
+from daqpytools.logging.handlerdataclasses import (
+    FormattedRichHandler,
+    HandlerType,
+    LogHandlerConf,
+)
+from daqpytools.logging.routing import (
+    AllowedHandlersStrategy,
+    StreamAwareAllowedHandlersStrategy,
+)
 from daqpytools.logging.utils import get_width
 
-
-class FormattedRichHandler(RichHandler):
-    """RichHandler that formats log messages with time, aligned columns, and styles."""
-
-    def __init__(self, width: int = 100) -> None:
-        """Initialize with custom console and style settings."""
-        console: Console = Console(
-            force_terminal=True, width=width, theme=CONSOLE_THEME
-        )
-        super().__init__(
-            console=console,
-            omit_repeated_times=False,
-            markup=True,
-            rich_tracebacks=True,
-            show_path=False,
-            show_time=False,  # We format time ourselves
-        )
-
-    def render(
-        self,
-        *,
-        record: logging.LogRecord,
-        traceback: object,
-        message_renderable: ConsoleRenderable,
-    ) -> Text:
-        """Render the log record into a rich Text object with custom formatting.
-
-        Args:
-            record (logging.LogRecord): The log record to render.
-            traceback (object): The traceback object (not used here).
-            message_renderable (ConsoleRenderable): The log message renderable.
-
-        Returns:
-            Text: The formatted log record as a rich Text object.
-
-        Raises:
-            None
-        """
-        dt: datetime = datetime.fromtimestamp(record.created, tz=TIME_ZONE)
-        padding: int = LOG_RECORD_PADDING.get("time", 25)
-        time_str: str = dt.strftime(DATE_TIME_FORMAT).ljust(padding)[:padding]
-        time_text: Text = Text(time_str, style="logging.time")
-
-        padding = LOG_RECORD_PADDING.get("level", 10)
-        level_text: Text = Text(
-            record.levelname.ljust(padding)[:padding],
-            style=self._get_level_style(record.levelno),
-        )
-
-        file_and_no: str = f"{record.filename}:{record.lineno}"
-        padding = LOG_RECORD_PADDING.get("file_and_line", 40)
-        file_and_no_text: Text = Text(
-            file_and_no.ljust(padding)[:padding], style="logging.location"
-        )
-
-        padding = LOG_RECORD_PADDING.get("logger_name", 45)
-        logger_name_text: Text = Text(
-            f"{record.name}".ljust(padding)[:padding], style="logging.logger_name"
-        )
-
-        # Convert message_renderable to Text for type consistency
-        message_text: Text
-        if isinstance(message_renderable, Text):
-            message_text = message_renderable
-        else:
-            message_text = Text.from_markup(str(message_renderable))
-
-        components: list[Text] = [
-            time_text,
-            level_text,
-            file_and_no_text,
-            logger_name_text,
-            message_text,
-        ]
-
-        return Text(" ").join(components)
-
-    def _get_level_style(self, level_no: int) -> str:
-        """Get the style string for the given log level number from the theme defined in
-        CONSOLE_THEME.
-
-        Args:
-            level_no (int): The log level number.
-
-        Returns:
-            str: The style string for the log level.
-        """
-        return str(
-            CONSOLE_THEME.styles.get(
-                f"logging.level.{logging_log_level_to_str(level_no).lower()}", ""
-            )
-        )
-
-# Initialise a logger to catch erstrace + other unknown handlertypes from OKS
-log: logging.Logger = logging.getLogger(__name__)
-log.addHandler(FormattedRichHandler(width=get_width()))
-log.setLevel("INFO")
-
-
-class StreamType(Enum):
-    """Enumtype to classify the set of relevant handlers (i.e streams)."""
-    BASE="base"
-    OPMON="opmon"
-    ERS="ers"
-
-@dataclass 
-class ProtobufConf:
-    """Dataclass to hold Protobuf Configuration."""
-    url:str= "monkafka.cern.ch"
-    port:int= 30092
-
-    def get_string(self) -> str:
-        """Converts back to string version."""
-        return f"{self.url}:{self.port}"
-
-class HandlerType(Enum):
-    """Enumtype to classify the existing set of Handlers.
-    Values must match exactly what is given in the OKS configuration, if any
-    All are in lowercase.
-    """
-    Unknown = "unknown"
-    Stream = "stream"
-    Rich = "rich"
-    File = "file"
-    Protobufstream = "protobufstream"
-    Lstdout = "lstdout"
-    Lstderr = "lstderr"
-    Throttle = "throttle"
-    @classmethod
-    def from_string(cls, s: str) -> HandlerType | None:
-        """Converts from a case-independent string to HandlerType."""
-        try:
-            return HandlerType(s.lower())
-        except ValueError:
-            msg=f"[red]{s}[/red] is not a known handler type"
-            log.warning(msg)
-            return None
-
-
-@dataclass
-class ERSPyLogHandlerConf:
-    """Dataclass that holds the relevant ERS configuration from OKS.
-
-    As an example, given the following
-    <obj class="Variable" id="ehn1-env-ers-error">
-        <attr name="name" type="string" val="DUNEDAQ_ERS_ERROR"/>
-        <attr name="value" type="string" val="erstrace,throttle,lstdout,
-            protobufstream(monkafka.cern.ch:30092)"/>
-    </obj>
-
-    This dataclass holds the list of relevant Handlers attached to the ERS log level. 
-    In case it also contains a single protobuf handler, the configuration is stored by 
-    the ProtobufConf instance. Multiple protobuf handlers with different url/ports 
-    are not yet supported.
-    """
-    handlers: list = field(default_factory = lambda: [])
-    protobufconf: ProtobufConf = field(default_factory = lambda: None)
-
-@dataclass
-class LogHandlerConf:
-    """Dataclass that holds the various streams and relevant handlers.
-    
-    Attributes:
-        init_ers: If True, automatically initializes ERS configuration 
-            during construction.
-        _BASE_HANDLERS: Private class variable for default base handlers
-        _OPMON_HANDLERS: Private class variable for opmon handlers
-        BASE_CONFIG: Class variable for base stream configuration
-        OPMON_CONFIG: Class variable for opmon stream configuration
-        ERS: Instance field for ERS configuration (loaded from environment)
-    """
-    init_ers: bool = False
-
-    _BASE_HANDLERS: ClassVar[set] = {HandlerType.Stream, HandlerType.Rich,
-        HandlerType.File
-        }
-    _OPMON_HANDLERS: ClassVar[set] = {HandlerType.Rich, HandlerType.Stream,
-        HandlerType.File}
-    _ERS: object = None
-    
-    Base: ClassVar[dict] = {
-        "handlers": _BASE_HANDLERS,
-        "stream": StreamType.BASE
-    }
-    
-    Opmon: ClassVar[dict] = {
-        "handlers": _OPMON_HANDLERS,
-        "stream": StreamType.OPMON
-    }
-
-    def __post_init__(self) -> None:
-        """Initialize ERS configuration if init_ers field is True.
-
-        This method is called automatically after dataclass initialization.
-        If the init_ers attribute was set to True, it triggers the complete
-        ERS initialization.
-        """
-        if self.init_ers:
-            self.init_ers_stream()
-
-    @property
-    def ERS(self) -> dict : # noqa: N802
-        """Get the ERS configuration dictionary.
-        
-        Returns:
-            dict: Contains 'ers_handlers' and 'stream' configuration for ERS
-            
-        Raises:
-            AttributeError: If ERS has not been
-                initialized (call init_ers_stream() first)
-        """
-        if not self._ERS:
-            err_msg = "ERS stream not initialised. Call init_ers_stream() first"
-            raise AttributeError(err_msg)
-        return self._ERS
-
-    def init_ers_stream(self) -> None:
-        """Initialize ERS configuration from environment variables.
-        
-        Loads ERS configuration from OKS environment variables and populates
-        the _ERS dict with handlers and stream information.
-        
-        Called automatically during construction if init_ers=True, or can be
-        called manually afterwards.
-        """
-        self._ERS = {
-            "ers_handlers":  LogHandlerConf._get_oks_conf(),
-            "stream": StreamType.ERS
-        }   
-
-    @staticmethod
-    def _convert_str_to_handlertype(handler_str: str) -> tuple[HandlerType,
-        ProtobufConf | None]:
-        """Parses a given environment variable to obtain the
-        HandlerType and ProtobufConf as necessary. 
-
-        Eg. converts "throttle" to HandlerType.Throttle
-            converts "protobufstream(url:port)" to return both the HandlerType and the 
-            protobuf configuration
-        """
-        if "erstrace" in handler_str:
-            msg = (
-                "ERSTrace is a C++ implementation, "
-                "does not have an equivalent in Python"
-            )
-            log.debug(msg)
-            return None, None
-
-        if HandlerType.Protobufstream.value not in handler_str:
-            return HandlerType.from_string(handler_str), None
-
-        match = re.search(r"\(([^:]+):(\d+)\)", handler_str)
-        if not match:
-            raise ProtobufFormatError(handler_str)
-        url, port = match.group(1), int(match.group(2))
-        return HandlerType.Protobufstream, ProtobufConf(url=url, port=port)
-
-    @staticmethod
-    def _make_ers_handler_conf(ers_log_level :str) -> ERSPyLogHandlerConf:
-        """Generates the ERSPyLogHandlerConf from reading an environment variable."""
-        erspyloghandlerconf = ERSPyLogHandlerConf()
-        envvalue = os.getenv(ers_log_level)
-        if envvalue is None:
-            raise ERSEnvError(ers_log_level)
-        
-        for h in envvalue.split(","):
-            handlertype, kafkaconf = LogHandlerConf._convert_str_to_handlertype(h)
-            erspyloghandlerconf.handlers.append(handlertype)
-            if kafkaconf:
-                erspyloghandlerconf.protobufconf = kafkaconf 
-        return erspyloghandlerconf
-
-    @staticmethod
-    def _get_oks_conf() -> dict:
-        """From the set of known environment variables, generate the ERS conf dict."""
-        return {var: LogHandlerConf._make_ers_handler_conf(var) 
-            for var in list(level_to_ers_var.values())}
-    
-    @staticmethod
-    def get_base() -> set[HandlerType]:
-        """Returns the default list of handlers from
-        LogHandlerConf._BASE_HANDLERS without having to initialise an instance.
-        """
-        return LogHandlerConf._BASE_HANDLERS
 
 class IssueRecord:
     """Tracks throttling state for a unique issue (identified by file: line)."""
@@ -341,37 +56,23 @@ class BaseHandlerFilter(logging.Filter):
     """Base filter that hold the logic on choosing if a handler should emit
     based on what HandlersTypes are supplied to it.
     """
-    def __init__(self, fallback_handlers: set[HandlerType] | None = None) -> None:
+    def __init__(
+        self,
+        fallback_handlers: set[HandlerType] | None = None,
+        allowed_handlers_strategy : AllowedHandlersStrategy | None = None,
+    ) -> None:
         """C'tor."""
-        if fallback_handlers is None:
-            fallback_handlers = set(LogHandlerConf.get_base())
-        self.fallback_handlers = fallback_handlers
+        self.fallback_handlers = set(fallback_handlers) if fallback_handlers is not None else LogHandlerConf.get_base() #! We should check if this is still the case
+
+        self.allowed_handlers_strategy = (
+            allowed_handlers_strategy or 
+            StreamAwareAllowedHandlersStrategy()
+        )
         super().__init__()
-    
-    def get_allowed(self, record: logging.LogRecord) -> list | None:
-        """Parses the record to obtain the set of Handlers that allows transmission."""
-        # TODO/future: kafkaprotobufs should validate url/port match before transmitting
-        
-        # Handle the ERS case, requires more processing
-        if getattr(record, "stream", None) == StreamType.ERS:
-            # Chain None checks using walrus operator
-            if (
-                # Checks if python log level has an ERS severity level match
-                (ers_level_var := level_to_ers_var.get(record.levelno)) is None
-                # Checks if ers_handlers (specifically for ERS) is supplied by the msg
-                or (ers_handlers := getattr(record, "ers_handlers", None)) is None
-                # _Assigns_ the relevant ERS handler conf from supplied level
-                or (ershandlerconf := ers_handlers.get(ers_level_var)) is None
-            ):
-                return None
-            
-            allowed = ershandlerconf.handlers
 
+    def get_allowed(self, record: logging.LogRecord) -> set[HandlerType] | None:
+        return self.allowed_handlers_strategy.resolve(record, self.fallback_handlers)
 
-        # Handle the non-ERS case
-        else:
-            allowed = getattr(record, "handlers", self.fallback_handlers) 
-        return allowed
         
 class HandleIDFilter(BaseHandlerFilter):
     """Filter class that accepts a list of 'allowed' handlers and will only fire
@@ -382,12 +83,14 @@ class HandleIDFilter(BaseHandlerFilter):
         self,
         handler_id: HandlerType | list[HandlerType],
         fallback_handlers: set[HandlerType] | None = None,
+        allowed_handlers_strategy: AllowedHandlersStrategy | None = None,
     ) -> None:
         """Initialises HandleIDFilter with the handler_id, to identify what
         kind of handler this filter is.
         """
         super().__init__(
-            fallback_handlers = fallback_handlers
+            fallback_handlers = fallback_handlers,
+            allowed_handlers_strategy=allowed_handlers_strategy
         )
         
         # Normalise handler_id to be a set
@@ -400,7 +103,7 @@ class HandleIDFilter(BaseHandlerFilter):
         """Identifies when a log message should be transmitted or not."""
         if not (allowed:= self.get_allowed(record)):
             return False
-        return bool(self.handler_ids & set(allowed))
+        return bool(self.handler_ids & allowed)
 
 class ThrottleFilter(BaseHandlerFilter):
     """Advanced logging filter with escalating throttle thresholds.
@@ -430,10 +133,12 @@ class ThrottleFilter(BaseHandlerFilter):
         fallback_handlers: set[HandlerType] | None = None,
         initial_threshold: int = 30,
         time_limit: int = 30,
+        allowed_handlers_strategy : AllowedHandlersStrategy | None = None
     ) -> None:
         """C'tor."""
         super().__init__(
-            fallback_handlers = fallback_handlers
+            fallback_handlers = fallback_handlers,
+            allowed_handlers_strategy = allowed_handlers_strategy
             )
         self.initial_threshold = initial_threshold
         self.time_limit = time_limit
